pyproject.toml
[tool.poetry]
name = "base_app"
version = "0.1.0"
description = ""
authors = ["Your Name Here"]

[tool.poetry.dependencies]
python = "^3.8"
celery = "4.4.5"
datadog = "^0.39.0"
ddtrace = "^0.45.0"
django = "3.1"
django-celery-beat = "2.0.0"
django-environ = "0.4.5"
django-extensions = "2.2.9"
django-filter = "2.3.0"
django-health-check = "3.12.1"
django-redis = "4.12.1"
django-storages = {extras = ["google"], version = "1.9.1"}
djangorestframework = "3.11.0"
drf-flex-fields = "0.8.5"
gevent = "^20.6.2"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.5"
python-json-logger = "^0.1.11"
pytz = "^2020.1"
redis = "^3.5.3"
sentry-sdk = "^0.16.1"
structlog = "^20.1.0"
uvicorn = "^0.11.5"

[tool.poetry.dev-dependencies]
bandit = "*"
django-stubs = "^1.5.0"
djangorestframework-stubs = "^1.2.0"
flake8 = "^3.8.3"
flake8-commas = "^2.0.0"
flake8-isort = "^3.0.1"
ipdb = "^0.13.2"
ipython = "^7.15.0"
isort = "^4.3.21"
mypy = "^0.770"
prospector = "^1.3.0"
pytest = "^5.4.3"
pytest-blockage = "^0.2.2"
pytest-cov = "^2.10.0"
pytest-django = "^3.9.0"
pytest-mock = "^3.1.1"
pytest-xdist = "^1.32.0"
werkzeug = "^1.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
